dryade-core/pyproject.toml
[project]
name = "dryade-core"
version = "1.0.0"
description = "Multi-agent orchestration platform"
authors = [{ name = "Dryade Team" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "crewai>=1.7.0",
    "pdfplumber>=0.10.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "email-validator>=2.0.0",
    "sqlalchemy>=2.0.25",
    "alembic>=1.13",
    "psycopg[binary]>=3.1",
    "redis>=5.0.0",
    "qdrant-client>=1.12.0",
    "httpx>=0.26.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "structlog>=24.1.0",
    "prometheus-client>=0.19.0",
    "passlib[argon2]>=1.7.4",
    "tenacity>=9.0.0",
    "py-cpuinfo>=9.0.0",
    "psutil>=7.0.0",
    "grpcio>=1.60.0",
    "circuitbreaker>=2.0.0",
    "sentence-transformers>=3.0.0",
    "apscheduler>=3.10.0",
    "pyotp>=2.9.0",
    "qrcode>=7.4.2",
    # Previously missing — hard unguarded imports in core/
    "PyYAML>=6.0",
    "PyJWT>=2.12.0",
    "cryptography>=42.0",
    "openai>=1.0",
    "anthropic>=0.20",
    "click>=8.0",
    "Jinja2>=3.1",
    "packaging>=20.0",
    "numpy>=1.21",
    # Agent framework integrations — required for full agent discovery
    "langchain-openai>=0.1",
    "langchain-anthropic>=0.1",
    "langgraph>=1.0.10",
    "google-adk>=0.1",
    "google-genai>=0.1",
    "google-generativeai>=0.1",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-exporter-otlp>=1.22.0",
]
mcp = [
    "mcp>=1.2.0",
]
docker = [
    "docker>=7.0",
]
litellm = [
    "litellm>=1.0",
]
aws = [
    "boto3>=1.28",
]
providers = [
    "cohere>=5.0",
    "mistralai>=1.0",
]
knowledge = [
    "fastembed>=0.1",
]
credentials = [
    "keyring>=24.0",
    "keyrings.cryptfile>=1.3",
]
neo4j = [
    "neo4j>=5.0",
]
security = [
    "pyClamd>=0.4",
]
pq = [
    "liboqs-python>=0.14.1",
]

[project.scripts]
dryade = "core.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "agents"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"

[dependency-groups]
dev = [
    "pytest>=9.0",
    "pytest-asyncio>=1.3",
    "pytest-timeout>=2.4",
    "pytest-cov>=5.0",
    "coverage[toml]>=7.0",
    "ruff==0.14.13",
]

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["core/ee/*.ee.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "T20", "W"]
ignore = ["E501", "N806", "N812", "N814", "N817", "N818", "F821", "E402"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["F841", "E741", "F811", "F401"]
"scripts/**" = ["F841", "T201", "E741"]
".sync/**" = ["T201"]
"core/ee/*.ee.py" = ["N999"]

[tool.coverage.run]
source = ["core"]
omit = [
    "core/alembic/*",
    "core/cli/*",
    "*/test_*.py",
    "*/__pycache__/*",
]

[tool.coverage.report]
fail_under = 15
show_missing = true
skip_empty = true
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
